fix: 프로필 이미지 없는 사용자 로그인 시 NPE 수정
