+++ backend/home-risk-check-spring/src/main/java/hanshin/home_risk_check/auth/controller/AuthController.java
@@ -4,9 +4,9 @@ import hanshin.home_risk_check.auth.dto.TokenResponse;
 import hanshin.home_risk_check.auth.service.AuthService;
 import hanshin.home_risk_check.auth.util.RefreshTokenCookieFactory;
 import hanshin.home_risk_check.global.dto.ApiResponse;
-import hanshin.home_risk_check.user.dto.LoginRequest;
-import hanshin.home_risk_check.user.dto.LoginResponse;
-import hanshin.home_risk_check.user.dto.SignupRequest;
+import hanshin.home_risk_check.auth.dto.LoginRequest;
+import hanshin.home_risk_check.auth.dto.LoginResponse;
+import hanshin.home_risk_check.auth.dto.SignupRequest;
 import hanshin.home_risk_check.user.dto.UserResponse;
 import hanshin.home_risk_check.user.entity.CustomUserDetails;
 import jakarta.validation.Valid;
