+++ backend/home-risk-check-spring/src/main/java/hanshin/home_risk_check/auth/dto/SignupRequest.java
@@ -1,11 +1,11 @@
-package hanshin.home_risk_check.user.dto;
+package hanshin.home_risk_check.auth.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 
-public record LoginRequest (
+public record SignupRequest (
 
     @NotBlank(message = "이메일은 비어 있을 수 없습니다.")
     @Email(message = "올바른 이메일 형식이 아닙니다.")
@@ -15,7 +15,11 @@ public record LoginRequest (
     @NotBlank(message = "비밀번호는 비어 있을 수 없습니다.")
     @Size(min = 8, max = 20, message = "비밀번호는 8~20자여야 합니다.")
     @Pattern(regexp = PASSWORD_PATTERN, message = "비밀번호는 영문, 숫자, 특수문자를 모두 포함해야 합니다.")
-    String password
+    String password,
+
+    @NotBlank(message = "닉네임은 비어 있을 수 없습니다.")
+    @Size(min = 2, max = 10, message = "닉네임은 2~10자여야 합니다.")
+    String nickname
 ) {
     private static final String PASSWORD_PATTERN =
             "^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)(?=.*[~!@#$%^&*()-_+=|{}\\[\\]:;<>.,?'\"])[A-Za-z\\d~!@#$%^&*()-_+=|{}\\[\\]:;<>.,?'\"]{8,20}$";
